WIP: implement start of fact-pipelining:
The idea is to speed up lots of similar facts (like a load of files).
Currently they are gathered one-by-one as required, however with larger
numbers of hosts all the channel creations add up. So this change will
enable pyinfra to "look ahead" in a deploy and generate a list of facts
it can pipeline (ie ls FILE; ls FILE, etc) via a single channel.

This step extracts the facts to gather and sets up the operation and
state handlers to ignore commands at this time. Upon completion the ops
are re-run with fact gathering enabled, same as before.

Next step is to implement the pipelining of facts in `pyinfra.api.facts`.