Fix silently including all hosts on typo
Making a typo in the name of a group, either in an inventory file
or in an include in deploy.py, results in all hosts being silently
ignored. This is due to the inventory returning None for unknown
group names.

This fixes the issue by not accepting unknown groups anymore.