Add support for pip3 or custom pip tool
Recent distributions such as Ubuntu 16.04 LTS do not ship with
Python2 by default. The `pip` binary available in the package
python3-pip is called `pip3`, and cannot be used currently by
pyinfra. This branch fixes the issue by allowing to specify
the name or path of the `pip` tool to use.