Ensure connections before running facts.
The upshot of this is that it enables running facts on hosts we're not actually deploying to. For example say we have a `master-01` and `slave-01` through `slave-04`. You might only want to deploy to the slaves (`--limit slave*`). This works fine until you need a fact from the master (`inventory.get_host('master-01').fact.X`) - this will now work as pyinfra will silently connect to the master upon the fact being requested.

This does mean that for any facts gathered after the master connects, they will be collected for the master also, but it's a small performance penalty and harmless (and might be possible to remove in the future!).