Implement `use_sudo_password` for the local + SSH connectors.
Both use roughly the same process:

+ Check if we've already got a sudo password, use that if so
+ Check if we've uploaded our `SUDO_ASKPASS` "exe", upload if needed
+ Use `SUDO_ASKASS` + environment variable to pass sudo password

This is secure _if_ `use_sudo_password=True`. It is also posssible to
explicitly set the password `use_sudo_password=pwd`, which possibly
leads to people sticking passwords in deploy code/etc. For now I shall
leave it as-is, because there are use-cases where it's better to explicitly
input the password (get from Vault, for example). Will add a warning to
the docs about this!