dockerssh: Add support for get/put file.
For get, generate a tempfile on the remote, tell the remote
docker to copy to the remote tempfile, then copy that
tempfile into to the local fs.

For put, generate a local tempfile, write the data into that,
copy it to the remote server, then remote execute docker and
tell it to run a docker cp.