Remove all fact (ab)use of `use_default_on_error`.
This would hide issues unrelated to the fact collection, such as auth
errors with sudo. The fact commands must now handle any expected failure
gracefully as part of the command. In most cases this simply means
checking for the existence of a binary with `which` and using a shell
or statement to execute `true` if not found, for example:

`dpkg -l` -> `which dpkg && dpkg -l || true`

This should properly bring out errors unrelated to facts whilst allowing
facts to be flexible (ie don't error a server if `dpkg` isn't present).