Enable passing a docker container ID directly.
This makes it possible to target an existing docker container from the
command line without using the `docker_container_id` host data variable.

Additionally this changes the disconnect behaviour to only commit and
stop the container when starting with an image. When executing against
a container we leave it as-is.