rework of the git.worktree operation
The previous `git.worktree` interface was not fully compatible with the
actual `git worktree` operation.

The new one is based on 3 things:
- a worktree path where to create the worktree,
- a list of options (`--detach`, `-b`, ...)
- a commitish parameter that represents a git element that can be used
  as a starting point to create the worktree (it could be a commit, a branch,
  or anything else).

This new `git.worktree` operation also pull the worktree when it is already present.
The pull operation is possible if a tracking branch is already configured for this
worktree or if a remote branch is provided thruogh the parameter `from_remote_branch`.