Remove the "will add user" detection for fact failure.
This was an attempt to detect the difference between a fact that failed
because a given user was missing but would be created by a previous
operation call. In reality there are way too many side effects by which
users get created (eg. installing system packages) that this detection
is ineffective, and we end up failing hosts when we need not.