Transfer ownership of files over ssh with sudo/su (#616)
* Transfer ownership of files over ssh with sudo/su.

Adds a step to SSH file uploads that assigns read permissions via
ACLs to the sudo/su user, and uses cp instead of mv to allow the
the intended user to take ownership.

* Use format to maintain python2 compatibility

* Modify StringCommand for consistent styling.

Co-authored-by: Nick Barrett <nick@fizzadar.com>

* Modify StringCommand for consistent styling.

Co-authored-by: Nick Barrett <nick@fizzadar.com>

* Only run setfacl command if a sudo/su user is specified

* Remove redundant chown command

The chown command is no longer required because we now cp the file as
the sudo/su user, so the file is created with the correct ownership.

* Remove temporary files after upload

* Modify test cases for new SSH upload commands

* Linting whitespaces.

* Expand SSH connector tests to cover copy & acl failures.

Co-authored-by: ben <ben@pop-os.localdomain>
Co-authored-by: Nick Barrett <nick@fizzadar.com>