Normalise base path for mocked local files.
This is a superficial patch to make sure internal paths to mocked files and directories are using the current OS format.
However, externally, especially in operation parameters, the paths remain in the user-provided format.
This is useful in tests because it means the same JSON can be used on Unix and Windows to describe tests.