Replace the deploy dir with the current working directory.
This removes the ambiguously defined / detected deploy directory
from which files/etc are loaded from. This means pyinfra will no
longer automatically pick a working directory. This is a good
thing as the detection mechanism was not consistent.

The next commit will implement a `--chdir` flag to set the working
directory before pyinfra executes.