Revise systemctl command options for facts.
The following flags were added or changed for both facts:
 --plain to ensure that status dots are not present in output
 --legend=false replaces the seemingly deprecated --no-legend flag

SystemdEnabled was not previously reporting statuses for instantiated
units because systemctl list-unit-files does not seem to output this
information consistently. This fact now uses systemctl list-units
instead to generate a list of units, and the enabled state of the
unit is obtained from systemctl show -P UnitFileState <unit> instead
of is-enabled. Upstream systemctl docs indicate that systemctl show
should be used when machine readable output is required.