Restore accessing facts & operations via `pyinfra.X.Y`
This restores automatic importing of all the facts and operations into
the `pyinfra.facts` & `pyinfra.operations` modules respectively, which
makes it possible to access them like `facts.server.X` and so on, without
importing the fact/op module directly.

This does come at a small cost in startup time which is a reasonable
tradeoff vs. flexible/expected import behaviour.