Implement nested operations
Nested operations are called during the execution phase within a callback
function passed into a `python.call` operation. Calling a nested operation
generates and immediately executes it on the target machine.

This is useful in complex scenarios where one operation output is
required in another. Nested operation calls should be kept to a minimunm
as they cannot be tracked as changes pre-execution.