Use unique SUDO_ASKPASS script per connection
Using a single SUDO_ASKPASS script prevents running multiple instances
of pyinfra against the same host, as one instance is likely to stomp on
the SUDO_ASKPASS while the other instance is still running.

We can fix the issue by running mktemp on the host to generate a unique
filename under `/tmp`. We chose `/tmp` in case the pyinfra connection
is interrupted before we remove the temp file, as it will get cleaned
out automatically by the host's normal processes for maintaining
temporary directories.

And as long as we're running mktemp on the host to generate the
temporary name, we may as well use that same shell command to generate
the SUDO_ASKPASS script itself, instead of sending it via SFTP. This
also allows us to work in scenarios where the host sshd doesn't accept
SFTP connections.