SUDO_ASKPASS_COMMAND: respect TMPDIR if it is set
Respect TMPDIR if it is set instead of forcing /tmp which on some hosts
may not be writable. Note that we're not using `-t` since it does not
behave in a portable way across platforms. e.g. macOS `-t` does NOT
want trailing XXXXXXX, while the BSDs do want them, and on Linux `-t`
is deprecated altogether.

Fixes #880.