Add `DEFAULT_TEMP_DIR` config variable, set `TEMP_DIR` to `None`
The latter will be used as a user-override, with the default `None` value
meaning "load temp dir from environment if possible". The new default value
provides a fallback if no environment variable is found.

This maintains compatibility with the current version - only change is
the default will now try to use any `TMPDIR` env on the remote host
which is an improvement.