Enable calling any function directly from the CLI
This removes the limitation on calling `<module>.<function>` with the
CLI and enables loading up any module/function, whether they are a
wrapped operation, deploy or just plain Python function.

Additionally this properly parallelises the generation of commands on
target hosts, previously for direct op calls this would be done in serial
unlike deploy files.

Also noting here that long term this may be a cleaner approach to using
pyinfra in general - always wrapping deploy code underneath a function
and calling it as a module rather than pointing at a .py file.