Update operation tests to remove all the complex idempotent checking
This is no longer necessary because operations don't modify or track
facts as there is no longer any fact cache at all. The new simplified
flow means we can just write simple tests with facts in the "correct"
state to test operations are idempotent.

This probably doesn't make a huge amount of sense and will be better
documented in the contributing guide.