Remove bare operation decorator style
The typed version of this does not work and seems to not be possible to
easily fix, there's zero benefit to the two forms of operation decorator
so removing the bare one keeps things simple and fixes the typing issue.

Also update op calls within ops to use kwargs to satisfy type checking.