Refactor and re-type operation decorator
Although this doesn't change how operations are used it is a significant
improvement (simplifcation) of the operation decorator and the typing of
it and the shared arguments available.

There is now an explicit `_inner` attribute operations must use when
calling operations in other operations, which hugely simplifies the
typing and actually makes it possible for mypy/pyright to correctly
type both user operation calls and internal calls.

Furthermore this inlines the operation typing removing the need for
a separate `operation.pyi` which is excluded from internal analysis.