Add `services` argument to systemd facts for filtering
This means the operation can now only request information about the
specific services targetted, thus avoiding fetching and parsing every
unit on the system.

This is especially important in v3 where facts don't get cached and
thus will be re-fetched every time the operation is called.

NOTE: the other system service facts/operations should also be updated
to behave like this, but since systemd is by far the most popular this
will yield the majority of the benefit immediately.