Replace pkg_resources.require() with a new implementation
pkg_resources is deprecated and, under some circumstances, causes import
errors in Python 3.12. This commit replaces the one remaining function
from pkg_resources, the require() function, with a new implementation
that does the same thing. It was necessary to implement the function
anew because there's no equivalent for require() in the packaging or
importlib.* modules. There is an implementation in the hbutils package,
but I didn't think it was worth bringing in a new dependency for just
this one function.
  https://hansbug.github.io/hbutils/main/api_doc/system/python.html#check-reqs
See also
  https://github.com/pypa/packaging-problems/issues/664
This new implementation is based most closely on the original
pkg_resources code.