`docker.container`: Refactor to support container recreation
This change refactors the way the `docker.container` operation manages
containers, in preparation for work to make recreation more intelligent.
This change is (mostly) a pure refactor; future changes will diff the
current container against the operation parameters to determine if a
container needs to be recreated. This will fix an issue where changing
any of the operation arguments does not result in actual container
changes upon execution.

In this refactor, container parameters are moved to a dedicated class,
which centralizes the `docker container` command-line argument
generation logic and the future diffing logic. The diff function is
roughed in, though it currently reports "no diff" (to match the
operation's current behavior). Since conditional recreation complicates
the operation's logic on which commands to execute, the decisions are
boosted into boolean variables to increase readability.

As a side benefit, supporting additional docker container parameters
should be more straightforward due to the centralization in said
dedicated class (I'm planning on adding support for container args, uid,
and other Docker params currently not supported).

The only behavioral change is that creating and starting a container is
no longer done in one exec (joined by `;`) but rather two separate
docker commands. This sidesteps questions about whether `;` is the
correct joiner (as opposed to `&&`) and reduces the amount of `kwargs`
fishing in the implementation.

Tested:
* `scripts/dev-test.sh` and `scripts/dev-test-e2e.sh` both pass, save
  for warnings also present prior to this change