`docker.container`: Support container args
This change adds an `args` parameter to the `docker.container` operation
that passes said supplied args to the container at creation time. This
allows the operation to support container images that have an entrypoint
expecting to receive additional arguments, without needing to build+push
a custom image that embeds said arguments.