[IMP] files.file: amend <touch> flag meaning
As it stands, the <touch> flag of `files.file` is useless: it only controls
whether a `touch <path>` command is issued when the <path> file already exists,
which results in a no-op, in both cases.

Since <assume_present> was removed in dfa69b66fe8ebf9c02accb1d3f6ec4905f11476d,
it may make sense to have <touch> control whether the operation is allowed to
create an empty file, if one is not found. And, raise if the file is not found
and the flag is unset: the usecase is that one may want to only change
ownership/permissions of an existing file, without accidentally creating an
empty placeholder for it.

A small unrelated change to the condition to set the mode is also included.