operations/server: improve support for mount operation in FreeBSD
The `mount` operation uses the `remount` option which is available
in linux distros, however in FreeBSD this option does not exist.
The equivalent is `update` which makes runtime changes to a mounted
file system. Not all file systems can change their options and not
all options can be changed, although this depends entirely on the
file system.